fix: refactor util file
