+++ src/tests/integration/omneo/benefit-definitions/update-benefit-definition.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, test, afterAll } from 'vitest'
 import { Omneo } from '../../../../omneo'
 import simpleOmneoRequest from '../../../lib/simple-omneo-request'
 import { BenefitDefinition } from '../../../../types'
-import { getRandomString } from './util'
+import { getRandomString } from '../../../lib/string/util'
 
 const omneo = new Omneo({
   tenant: process.env.OMNEO_TENANT as string,
