+++ src/tests/integration/omneo/countries/delete-country.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, test, afterAll } from 'vitest'
 import { Omneo } from '../../../../omneo'
 import simpleOmneoRequest from '../../../lib/simple-omneo-request'
 import { CountryInput } from '../../../../types'
-import { getName, getIso2, getIso3, getIsoNumeric } from './util'
+import { getRandomString, getIsoNumeric } from '../../../lib/string/util'
 
 const omneo = new Omneo({
   tenant: process.env.OMNEO_TENANT as string,
@@ -13,9 +13,9 @@ const FAILED_DELETE_COUNTRIES : number[] = []
 describe('Countries delete', () => {
   test('SDK Countries delete', async () => {
     const payload: CountryInput = {
-      name: getName(),
-      iso_2: getIso2(),
-      iso_3: getIso3(),
+      name: getRandomString('sdk_unit_test_country'),
+      iso_2: getRandomString('sdk_unit_test_iso2'),
+      iso_3: getRandomString('sdk_unit_test_iso3'),
       iso_numeric: getIsoNumeric(),
       sort_order: null
     }
