+++ src/tests/integration/omneo/profiles/attributes/custom/list-profile-custom-attirbutes.test.ts
@@ -2,7 +2,7 @@ import { describe, test, expect, afterAll } from 'vitest'
 import { Omneo } from '../../../../../../omneo'
 import simpleOmneoRequest from '../../../../../lib/simple-omneo-request'
 import { CustomAttribute } from '../../../../../../types'
-import { getRandomString } from './util'
+import { getRandomString } from '../../../../../lib/string/util'
 
 const omneo = new Omneo({
   tenant: process.env.OMNEO_TENANT as string,
