+++ src/tests/integration/omneo/profiles/attributes/custom/update-profile-custom-attirbute.test.ts
@@ -1,7 +1,7 @@
 import { describe, test, expect, afterAll } from 'vitest'
 import { Omneo } from '../../../../../../omneo'
 import simpleOmneoRequest from '../../../../../lib/simple-omneo-request'
-import { getRandomString } from './util'
+import { getRandomString } from '../../../../../lib/string/util'
 import { CustomAttribute } from '../../../../../../types'
 
 const omneo = new Omneo({
