--- src/tests/integration/omneo/profiles/attributes/custom/util.ts
@@ -1,2 +1,0 @@
-import randomString from '../../../../../lib/string/random'
-export const getRandomString = (name: string) => { return `${name}_${randomString(5).toLowerCase()}_${Math.floor(Date.now() / 1000)}` }
