--- src/tests/integration/omneo/reward-definitions/util.ts
@@ -1,3 +1,0 @@
-import randomString from '../../../lib/string/random'
-
-export const getRandomString = (name: string) => { return `${name}_${randomString(5).toLowerCase()}_${Math.floor(Date.now() / 1000)}` }
