+++ src/tests/integration/omneo/rewards/get-reward.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, test, beforeAll, afterAll } from 'vitest'
 import { Omneo } from '../../../../omneo'
 import simpleOmneoRequest from '../../../lib/simple-omneo-request'
 import { Reward, RewardDefinitionCreateInput } from '../../../../types'
-import { getRandomString, convertToUTC } from './util'
+import { getRandomString, convertToUTC } from '../../../lib/string/util'
 
 const omneo = new Omneo({
   tenant: process.env.OMNEO_TENANT as string,
