+++ src/tests/integration/omneo/rewards/update-reward.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, test, beforeAll, afterAll } from 'vitest'
 import { Omneo } from '../../../../omneo'
 import simpleOmneoRequest from '../../../lib/simple-omneo-request'
 import { Reward, RewardDefinitionCreateInput, RewardUpdateInput } from '../../../../types'
-import { getRandomString, convertToUTC } from './util'
+import { getRandomString, convertToUTC } from '../../../lib/string/util'
 
 const omneo = new Omneo({
   tenant: process.env.OMNEO_TENANT as string,
